% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{autoplot.spline}
\alias{autoplot.spline}
\title{Creates \link[ggplot2]{ggplot} of \link[stats]{spline} Object}
\usage{
\method{autoplot}{spline}(object, ...)
}
\arguments{
\item{object}{a \link[stats]{spline} object}

\item{...}{..}
}
\value{
Function \link{autoplot.spline} returns a \link[ggplot2]{ggplot} figure.
}
\description{
..
}
\note{
Function \link{autoplot.spline} is not substantially different from \code{splines:::plot.spline}.
}
\examples{
class(x1 <- splines::interpSpline(obj1 = weight ~ height, obj2 = women, bSpline = TRUE))
plot(x1)
autoplot(x1)

class(x2 <- splines::polySpline(x1))
plot(x2)
autoplot(x2)
}
\keyword{internal}
