% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBar.R
\name{ggBar}
\alias{ggBar}
\title{ggBar}
\usage{
ggBar(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{\link[base]{factor} \link[base]{matrix}}

\item{na.rm}{\link[base]{logical} scalar}

\item{...}{..}
}
\description{
ggBar
}
\examples{
x = sample.int(4L, size = 100L, replace = TRUE) |> 
  array(dim = c(20L, 5L), dimnames = list(NULL, LETTERS[1:5]))
attr(x, 'levels') = letters[1:4]
class(x) = 'factor'
head(x)
# ggBar(x, position = 'dodge', colour = 'white') # move to vignette

}
