% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBoxJitter.R
\name{ggBoxJitter}
\alias{ggBoxJitter}
\title{ggBoxJitter}
\usage{
ggBoxJitter(
  data,
  y,
  x,
  colour = NULL,
  violin = FALSE,
  dodge.width = 0.5,
  data.name = deparse1(substitute(data)),
  htest,
  ...
)
}
\arguments{
\item{data}{..}

\item{x, y, colour}{..}

\item{violin}{..}

\item{dodge.width}{..}

\item{data.name}{..}

\item{...}{..}
}
\description{
..
}
\details{
Stop doing \link[ggplot2]{geom_violin}, need to educate people what they are.
}
\examples{
ggBoxJitter(data = CO2, y = uptake, x = Type)
ggBoxJitter(data = CO2, y = uptake, x = Treatment, colour = Type)
ggBoxJitter(data = CO2, y = uptake, x = Plant, colour = Type, violin = TRUE) + 
  facet_grid(rows = vars(Treatment))
}
\keyword{internal}
