% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg2D.R
\name{ggContour}
\alias{ggContour}
\title{ggContour}
\usage{
ggContour(data, z, x, y, ...)
}
\arguments{
\item{data}{..}

\item{x, y, z}{..}

\item{...}{..}
}
\description{
..
}
\details{
Function \link{ggContour} improves \link[graphics]{contour}
}
\examples{
?volcano
graphics::filled.contour(volcano)
ggContour(volcano)

}
