% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg2D.R
\name{ggHeatmap}
\alias{ggHeatmap}
\title{ggHeatmap}
\usage{
ggHeatmap(data, fill, x, y, ...)
}
\arguments{
\item{data}{..}

\item{fill}{..}

\item{x, y}{..}

\item{...}{..}
}
\description{
..
}
\examples{
ggHeatmap(VADeaths)
ggHeatmap(log1p(occupationalStatus))

}
