% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggHist.R
\name{ggHist}
\alias{ggHist}
\title{ggHist}
\usage{
ggHist(object, rel = FALSE, by_integer = FALSE, ...)
}
\arguments{
\item{object}{..}

\item{rel}{\link[base]{logical} scalar, whether a histogram with relative frequencies should be
produced (default \code{FALSE})}

\item{by_integer}{..}

\item{...}{..}
}
\description{
..
}
\details{
Function \link{ggHist} is the \CRANpkg{ggplot2} version of \link[graphics]{hist}
}
\examples{
\dontrun{
ggHist(stack.loss)
ggHist(stack.loss, rel = TRUE)
library(plotly) # https://plotly.com/r/histograms/
plot_ly(x = ~stack.loss, type = "histogram")
plot_ly(x = ~stack.loss, type = "histogram", histnorm = "probability")

# 'Date' object

EXP_Date1 = structure(c(15863, 15887, NA, 15873, 15885, 15878), class = 'Date')
EXP_Date2 = structure(c(15965, 15917, 15923, 15955, 15918, NA), class = 'Date')
stopifnot(is.double(EXP_Date1), !is.numeric(EXP_Date1))
print.default(EXP_Date1)
ggHist(EXP_Date1)
# ggBoxJitter(list(EXP_Date1, EXP_Date2))

# 'POSIXt' object

EXP_ct1 = c(Sys.time() + 3600 * rnorm(20))
typeof(EXP_ct1)
stopifnot(is.double(EXP_ct1), !is.numeric(EXP_ct1))
EXP_ct2 = c(Sys.time() + 3600 * rnorm(20, mean = 5))
EXP_lt1 = as.POSIXlt.POSIXct(EXP_ct1)
EXP_lt2 = as.POSIXlt.POSIXct(EXP_ct2)
stopifnot(!is.double(EXP_lt2), !is.numeric(EXP_lt2))
typeof(EXP_lt1) # list
print.default(EXP_lt2) # wow!
ggHist(EXP_ct1)
ggHist(EXP_lt2)
#ggBoxJitter(list(EXP_ct1, EXP_ct2))
#ggBoxJitter(list(as.POSIXlt(EXP_ct1), EXP_lt2))
#ggBoxJitter(list(EXP_ct1, c(Sys.time() + 3600 * rnorm(20, mean = 5))))

# 'difftime' object

EXP_dft = EXP_Date2 - EXP_Date1
stopifnot(is.double(EXP_dft), !is.numeric(EXP_dft))
ggHist(EXP_dft)
}
}
