% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{ggplot.tzh-package}
\alias{ggplot.tzh}
\alias{ggplot.tzh-package}
\title{ggplot.tzh}
\description{
..
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

}
