% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggQQ.R
\name{ggQQ}
\alias{ggQQ}
\title{ggQQ}
\usage{
ggQQ(object, fun = qnorm, ...)
}
\arguments{
\item{object}{..}

\item{fun}{..}

\item{...}{..}
}
\description{
..
}
\details{
Function \link{ggQQ} improves \link[stats]{qqnorm} and \link[stats]{qqplot}
}
\examples{
ggQQ(precip, mean = mean.default(precip), sd = sd(precip))

}
\references{
\url{http://stackoverflow.com/questions/4357031/qqnorm-and-qqline-in-ggplot2}
}
