% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggScatter.R
\name{ggScatter}
\alias{ggScatter}
\alias{ggScatter2}
\title{Scatter Plots using \CRANpkg{ggplot2}}
\usage{
ggScatter(
  data,
  y,
  x,
  colour = NULL,
  shape = NULL,
  jitter = 0,
  data.name = deparse1(substitute(data)),
  ...
)

ggScatter2(data, y1, y2, x, jitter = 0, ...)
}
\arguments{
\item{data}{\link[base]{data.frame} or \link[base]{list}}

\item{x, y, y1, y2}{\link[base]{expression}s, currently only \link[base]{name}s are tested}

\item{colour, shape}{(optional) \link[base]{expression}s, currently only \link[base]{name}s are tested}

\item{jitter, ...}{additional parameters of \link[ggplot2]{geom_jitter}}

\item{data.name}{..}
}
\description{
..
}
\details{
Function \code{\link[=ggScatter]{ggScatter()}} produces a scatter plot.

Function \link{ggScatter2} produces a scatter plot with two vertical axes.
}
\note{
Potential name clash with \code{ggpubr::ggscatter}.
}
\examples{
ggScatter(iris, x = Sepal.Length, y = Petal.Length)
ggScatter(iris, x = Sepal.Length, y = Petal.Length, colour = Species)
library(plotly)
plot_ly(iris, x = ~Sepal.Length, y = ~Petal.Length, type = 'scatter', mode = 'markers')
plot_ly(iris, x = ~Sepal.Length, y = ~Petal.Length, color = ~Species, 
  type = 'scatter', mode = 'markers')

car = mtcars |> within(expr = {
 cyl = factor(cyl)
 vs = as.logical(vs)
})
ggScatter(car, x = mpg, y = drat, colour = cyl, shape = vs) + facet_grid(cols = vars(gear))
# not sure how to facet plotly
ggScatter2(swiss, y1 = Fertility, y2 = Agriculture, x = Infant.Mortality)

}
